move legacy stuff to a separate folder and modules
