+++ merino/curated_recommendations/__init__.py
@@ -22,6 +22,7 @@ from merino.curated_recommendations.prior_backends.gcs_prior import GcsPrior
 from merino.curated_recommendations.prior_backends.constant_prior import ConstantPrior
 from merino.curated_recommendations.prior_backends.protocol import PriorBackend
 from merino.curated_recommendations.provider import CuratedRecommendationsProvider
+from merino.curated_recommendations.legacy.provider import LegacyCuratedRecommendationsProvider
 from merino.utils.metrics import get_metrics_client
 from merino.utils.http_client import create_http_client
 from merino.utils.synced_gcs_blob import SyncedGcsBlob
@@ -32,6 +33,7 @@ from merino.utils.gcs import initialize_storage_client
 logger = logging.getLogger(__name__)
 
 _provider: CuratedRecommendationsProvider
+_legacy_provider: LegacyCuratedRecommendationsProvider
 
 
 def init_engagement_backend() -> EngagementBackend:
@@ -91,6 +93,7 @@ def init_prior_backend() -> PriorBackend:
 def init_provider() -> None:
     """Initialize the curated recommendations provider."""
     global _provider
+    global _legacy_provider
 
     engagement_backend = init_engagement_backend()
 
@@ -114,9 +117,21 @@ def init_provider() -> None:
         prior_backend=init_prior_backend(),
         sections_backend=sections_backend,
     )
+    _legacy_provider = LegacyCuratedRecommendationsProvider(
+        scheduled_surface_backend=scheduled_surface_backend,
+        engagement_backend=engagement_backend,
+        prior_backend=init_prior_backend(),
+        sections_backend=sections_backend,
+    )
 
 
 def get_provider() -> CuratedRecommendationsProvider:
     """Return the curated recommendations provider."""
     global _provider
     return _provider
+
+
+def get_legacy_provider() -> LegacyCuratedRecommendationsProvider:
+    """TODO"""
+    global _legacy_provider
+    return _legacy_provider
