+++ merino/curated_recommendations/utils.py
@@ -3,15 +3,10 @@
 import re
 import time
 
-from urllib.parse import quote
-from pydantic import HttpUrl
 from merino.curated_recommendations.corpus_backends.protocol import SurfaceId
 from merino.curated_recommendations.protocol import (
     CuratedRecommendationsRequest,
     Locale,
-    CuratedRecommendation,
-    CuratedRecommendationLegacy,
-    CuratedRecommendationGlobalLegacy,
 )
 
 
@@ -95,49 +90,3 @@ def is_enrolled_in_experiment(
 def get_millisecond_epoch_time() -> int:
     """Return the time in milliseconds since the epoch as an integer."""
     return int(time.time() * 1000)
-
-
-def transform_image_url_to_pocket_cdn(original_url: HttpUrl) -> HttpUrl:
-    """Transform an original image URL to a Pocket CDN URL for the given image with a fixed width of 450px.
-
-    The original URL is encoded and embedded as a query parameter.
-    """
-    encoded_url = quote(str(original_url), safe="")
-    return HttpUrl(f"https://img-getpocket.cdn.mozilla.net/direct?url={encoded_url}&resize=w450")
-
-
-def map_curated_recommendations_to_legacy_recommendations(
-    base_recommendations: list[CuratedRecommendation],
-) -> list[CuratedRecommendationLegacy]:
-    """Map CuratedRecommendation object to CuratedRecommendationDesktopLegacy"""
-    return [
-        CuratedRecommendationLegacy(
-            typename="Recommendation",
-            recommendationId=item.corpusItemId,
-            tileId=item.tileId,
-            url=item.url,
-            title=item.title,
-            excerpt=item.excerpt,
-            publisher=item.publisher,
-            imageUrl=item.imageUrl,
-        )
-        for item in base_recommendations
-    ]
-
-
-def map_curated_recommendations_to_legacy_global_recommendations(
-    base_recommendations: list[CuratedRecommendation],
-) -> list[CuratedRecommendationGlobalLegacy]:
-    """Map CuratedRecommendation object to CuratedRecommendationGlobalLegacy"""
-    return [
-        CuratedRecommendationGlobalLegacy(
-            id=item.tileId,
-            title=item.title,
-            url=item.url,
-            excerpt=item.excerpt,
-            domain=item.publisher,
-            image_src=transform_image_url_to_pocket_cdn(item.imageUrl),
-            raw_image_src=item.imageUrl,
-        )
-        for item in base_recommendations
-    ]
